refactor: 테스트에서만 사용되는 메서드 제거
