+++ src/test/java/domain/RefereeTest.java
@@ -1,14 +1,5 @@
 package domain;
 
-import domain.Card;
-import domain.Cards;
-import domain.Dealer;
-import domain.Denomination;
-import domain.Gambler;
-import domain.Player;
-import domain.PlayerName;
-import domain.Suit;
-import game.Referee;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
